+++ backend/internal/handlers/auth/types.go
@@ -1,9 +1,26 @@
 package auth
 
+import "go.mongodb.org/mongo-driver/mongo"
+
+type Service struct {
+	users *mongo.Collection
+}
+
+func newService(collections map[string]*mongo.Collection) *Service {
+	return &Service{collections["users"]}
+}
+type Handler struct {
+	service *Service
+}
+type TokenResponse struct {
+	AccessToken string `json:"access_token"`
+	RefreshToken   string `json:"refresh_token"`
+	User        string `json:"user"`
+}
+
 type User struct {
 	Email string `bson:"email"`
 }
-
 type LoginRequest struct {
 	Email    string `json:"email"`
 	Password string `json:"password"`
